Move APR calcs to one place
